Improve FieldFormula unit test.
- merge field_formula_test and field_evaluate_formula_test
- remove tests of value method, use only new assemblation
